Rearrange ChangeDrafts directory
Rearrange the contents of the ChangeDrafts directory as follows:

- Move drafts that have been accepted to the Accepted subdirectory.
- Move drafts that have been rejected to the Rejected subdirectory.
- Move drafts that are currently being authored or are still
    under discussion to the Active subdirectory.

This will make it easier to see the issues we're still working on,
and it will clarify the disposition of resolved issues.

While performing this exercise, I noticed that I'd neglected to merge the
EXAMPLE in Section 3.7.2.3 (Tag metadata) from the change draft to the
provisional draft, so I fixed that as well.