Add sarif-schema.json
Add the SARIF schema file as it currently exists on the sarif-v2 branch of
the Microsoft/sarif-sdk repo. This version of the schema has the SARIF v2
changes mentioned in "Schema changes between SARIF v1 and v2.md", as they
appear in order in that file, from the beginning of the file through the
changes for Issue #82, "Add instance id to result object".

I'll make a separate commit to incorporate the changes for each
subsequent issue mentioned in the "Schema changes" document.