Revert change to CSD.1 schema
A previous commit removed a line from the CSD.1 version of the schema.
The line that was removed was incorrect, and deserved to be removed,
but we can't touch the CSD.1 schema, which has already been published
and commented on.

Instead, we will record it as an Erratum, and correct it in CSD.2. In fact,
the same commit that touched the CSD.1 version of the schema made
the fix in the "active" version of the schema.