#179: Don't replace conversion.tool with conversion.driver
conversion.tool is the conversion scenario analog of run.tool.
Both the direct production and the conversion scenarios are
accomplished by a tool that consists of a driver, and possibly
a set of auxiliary files including plugins and configuration files.
In the conversion scenario, configuration files are more likely
than plugins, but that’s neither here nor there.