Change draft for #321: Fix bug related to externalProperties.$schema
The change draft said that if an externalProperties object occurs as an
element of sarifLog.inlineExternalProperties, then if its $schema
property is omitted, it inherits the value from the containing
sarifLog.$schema. But this is wrong, because sarifLog.$schema is the
schema for a SARIF log file, while externalProperties.$schema is the
schema for an external properties file.

The change draft correctly said that externalProperties.version can
inherit from sarifLog.version. This is right because both properties
refer to a version of the SARIF specification.